feat(api): add is_persistent field to personal access tokens

+++ app/Http/Controllers/Api/AppTokenController.php
@@ -17,7 +17,10 @@ class AppTokenController extends Controller
         }
 
         DB::beginTransaction();
-        $token->delete();
+
+        if (!(bool)$token->is_persistent) {
+            $token->delete();
+        }
 
         $userAgent = explode('/', $request->header('User-Agent'));
         $token = $token->tokenable->createToken($userAgent[0], expiresAt: now()->addMonths(3));
